requirements.txt
numpy
pandas
librosa
praat-parselmouth
soundfile
